use integer for keys, attempt travis fix

+++ server/models.py
@@ -89,7 +89,7 @@ class Model(db.Model):
 
 
 class User(Model, UserMixin):
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255))
     email = db.Column(db.String(255, collation='utf8_bin'),
         unique=True, nullable=False, index=True)
@@ -120,7 +120,7 @@ class User(Model, UserMixin):
 
 
 class Course(Model):
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     # offering - E.g., 'cal/cs61a/fa14'
     offering = db.Column(db.String(255), nullable=False, unique=True, index=True)
     institution = db.Column(db.String(255), nullable=False)  # E.g., 'UC Berkeley'
@@ -152,7 +152,7 @@ class Assignment(Model):
         url - cs61a.org/proj/hog/hog.zip
     """
 
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255), index=True, nullable=False, unique=True)
     course_id = db.Column(db.ForeignKey("course.id"), index=True,
                           nullable=False)
@@ -345,7 +345,7 @@ class Message(Model):
     __tablename__ = 'message'
     __table_args__ = {'mysql_row_format': 'COMPRESSED'}
 
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     backup_id = db.Column(db.ForeignKey("backup.id"), nullable=False, index=True)
     contents = db.Column(JsonBlob, nullable=False)
     kind = db.Column(db.String(255), nullable=False, index=True)
@@ -354,7 +354,7 @@ class Message(Model):
 
 
 class Backup(Model):
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
 
     submitter_id = db.Column(db.ForeignKey("user.id"), nullable=False)
     assignment_id = db.Column(db.ForeignKey("assignment.id"), nullable=False)
@@ -440,7 +440,7 @@ class Group(Model):
     A group must have at least 2 participants.
     Degenerate groups are deleted.
     """
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     assignment_id = db.Column(db.ForeignKey("assignment.id"), nullable=False)
 
     assignment = db.relationship("Assignment")
@@ -601,7 +601,7 @@ class GroupAction(Model):
     """A group event, for auditing purposes. All group activity is logged."""
     action_types = ['invite', 'accept', 'decline', 'remove']
 
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     action_type = db.Column(db.Enum(*action_types, name='action_type'), nullable=False)
     # user who initiated request
     user_id = db.Column(db.ForeignKey("user.id"), nullable=False)
